+++ src/testing/mod.rs
@@ -3,4 +3,5 @@
 mod fake_server;
 
 pub use self::fake_server::FakeServer;
+pub use document::DocumentBuilder;
 pub use view::{IsGrouped, IsReduced, IsUnreduced, ViewResponseBuilder};